requirements.txt
# AI Services
openai>=1.0.0
azure-ai-inference>=1.0.0b1

# File format readers
ebooklib>=0.18
mobi>=0.3.3
python-docx>=1.0.0
markdown>=3.5
beautifulsoup4>=4.12.0

# Utilities
python-dotenv>=1.0.0
